requirements.txt
asgiref==3.3.1
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.6
django-cors-headers==3.7.0
django-nose==1.4.7
djangorestframework==3.12.2
gunicorn==20.0.4
httpretty==1.0.5
nose==1.3.7
Pillow==8.1.0
pinocchio==0.4.2
psycopg2-binary==2.8.6
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
whitenoise==5.2.0